+++ apps/app/src/server/routes/admin.js
@@ -7,6 +7,7 @@ import { exportService } from '../service/export';
 const logger = loggerFactory('growi:routes:admin');
 
 /* eslint-disable no-use-before-define */
+/** @param {import('~/server/crowi').default} crowi Crowi instance */
 module.exports = function(crowi, app) {
   const ApiResponse = require('../util/apiResponse');
   const importer = require('../util/importer')(crowi);
@@ -103,7 +104,7 @@ module.exports = function(crowi, app) {
       return res.json(ApiResponse.error('esa.io form is blank'));
     }
 
-    await configManager.updateConfigsInTheSameNamespace('crowi', form);
+    await configManager.updateConfigs(form);
     importer.initializeEsaClient(); // let it run in the back aftert res
     const parameters = { action: SupportedAction.ACTION_ADMIN_ESA_DATA_UPDATED };
     activityEvent.emit('update', res.locals.activity._id, parameters);
@@ -125,7 +126,7 @@ module.exports = function(crowi, app) {
       return res.json(ApiResponse.error('Qiita form is blank'));
     }
 
-    await configManager.updateConfigsInTheSameNamespace('crowi', form);
+    await configManager.updateConfigs(form);
     importer.initializeQiitaClient(); // let it run in the back aftert res
     const parameters = { action: SupportedAction.ACTION_ADMIN_QIITA_DATA_UPDATED };
     activityEvent.emit('update', res.locals.activity._id, parameters);