+++ apps/app/src/server/routes/apiv3/export.js
@@ -116,7 +116,7 @@ const router = express.Router();
  *             type: boolean
  *             description: whether the current exporting job exists or not
  */
-
+/** @param {import('~/server/crowi').default} crowi Crowi instance */
 module.exports = (crowi) => {
   const loginRequired = require('../../middlewares/login-required')(crowi);
   const adminRequired = require('../../middlewares/admin-required')(crowi);