+++ apps/app/src/server/routes/apiv3/search.js
@@ -17,6 +17,7 @@ const router = express.Router();
 
 const noCache = require('nocache');
 
+/** @param {import('~/server/crowi').default} crowi Crowi instance */
 module.exports = (crowi) => {
   const loginRequired = require('../../middlewares/login-required')(crowi);
   const adminRequired = require('../../middlewares/admin-required')(crowi);