+++ apps/app/src/server/service/file-uploader/gcs/index.ts
@@ -23,14 +23,18 @@ import { GcsMultipartUploader } from './multipart-uploader';
 const logger = loggerFactory('growi:service:fileUploaderGcs');
 
 
-function getGcsBucket() {
-  return configManager.getConfig('crowi', 'gcs:bucket');
+function getGcsBucket(): string {
+  const gcsBucket = configManager.getConfig('gcs:bucket');
+  if (gcsBucket == null) {
+    throw new Error('GCS bucket is not configured.');
+  }
+  return gcsBucket;
 }
 
 let storage: Storage;
 function getGcsInstance() {
   if (storage == null) {
-    const keyFilename = configManager.getConfig('crowi', 'gcs:apiKeyJsonPath');
+    const keyFilename = configManager.getConfig('gcs:apiKeyJsonPath');
     // see https://googleapis.dev/nodejs/storage/latest/Storage.html
     storage = keyFilename != null
       ? new Storage({ keyFilename }) // Create a client with explicit credentials
@@ -40,8 +44,7 @@ function getGcsInstance() {
 }
 
 function getFilePathOnStorage(attachment: IAttachmentDocument) {
-  const namespace = configManager.getConfig('crowi', 'gcs:uploadNamespace');
-  // const namespace = null;
+  const namespace = configManager.getConfig('gcs:uploadNamespace');
   let dirName: string;
   if (attachment.attachmentType === AttachmentType.PAGE_BULK_EXPORT) {
     dirName = FilePathOnStoragePrefix.pageBulkExport;
@@ -52,7 +55,7 @@ function getFilePathOnStorage(attachment: IAttachmentDocument) {
   else {
     dirName = FilePathOnStoragePrefix.user;
   }
-  const filePath = urljoin(namespace || '', dirName, attachment.fileName);
+  const filePath = urljoin(namespace, dirName, attachment.fileName);
 
   return filePath;
 }
@@ -74,7 +77,14 @@ class GcsFileUploader extends AbstractFileUploader {
    * @inheritdoc
    */
   override isValidUploadSettings(): boolean {
-    throw new Error('Method not implemented.');
+    try {
+      getGcsBucket();
+      return true;
+    }
+    catch (err) {
+      logger.error(err);
+      return false;
+    }
   }
 
   /**
@@ -102,7 +112,7 @@ class GcsFileUploader extends AbstractFileUploader {
    * @inheritdoc
    */
   override determineResponseMode() {
-    return configManager.getConfig('crowi', 'gcs:referenceFileWithRelayMode')
+    return configManager.getConfig('gcs:referenceFileWithRelayMode')
       ? ResponseMode.RELAY
       : ResponseMode.REDIRECT;
   }
@@ -177,7 +187,7 @@ class GcsFileUploader extends AbstractFileUploader {
     const myBucket = gcs.bucket(getGcsBucket());
     const filePath = getFilePathOnStorage(attachment);
     const file = myBucket.file(filePath);
-    const lifetimeSecForTemporaryUrl = configManager.getConfig('crowi', 'gcs:lifetimeSecForTemporaryUrl');
+    const lifetimeSecForTemporaryUrl = configManager.getConfig('gcs:lifetimeSecForTemporaryUrl');
 
     // issue signed url (default: expires 120 seconds)
     // https://cloud.google.com/storage/docs/access-control/signed-urls
@@ -224,8 +234,8 @@ module.exports = function(crowi: Crowi) {
   const lib = new GcsFileUploader(crowi);
 
   lib.isValidUploadSettings = function() {
-    return configManager.getConfig('crowi', 'gcs:apiKeyJsonPath') != null
-      && configManager.getConfig('crowi', 'gcs:bucket') != null;
+    return configManager.getConfig('gcs:apiKeyJsonPath') != null
+      && configManager.getConfig('gcs:bucket') != null;
   };
 
   (lib as any).deleteFile = function(attachment) {
@@ -271,8 +281,8 @@ module.exports = function(crowi: Crowi) {
    * - per-file size limit (specified by MAX_FILE_SIZE)
    */
   (lib as any).checkLimit = async function(uploadFileSize) {
-    const maxFileSize = configManager.getConfig('crowi', 'app:maxFileSize');
-    const gcsTotalLimit = configManager.getConfig('crowi', 'app:fileUploadTotalLimit');
+    const maxFileSize = configManager.getConfig('app:maxFileSize');
+    const gcsTotalLimit = configManager.getConfig('app:fileUploadTotalLimit');
     return lib.doCheckLimit(uploadFileSize, maxFileSize, gcsTotalLimit);
   };
 
@@ -287,7 +297,7 @@ module.exports = function(crowi: Crowi) {
     const gcs = getGcsInstance();
     const bucket = gcs.bucket(getGcsBucket());
     const [files] = await bucket.getFiles({
-      prefix: configManager.getConfig('crowi', 'gcs:uploadNamespace'),
+      prefix: configManager.getConfig('gcs:uploadNamespace'),
     });
 
     return files.map(({ name, metadata: { size } }) => {